+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Released: Upcoming
 - Added file handling events
 - Added some more IDs and classes in the HTML to be more tweakable
 - Fixed a crash if the table name of an entity is a MySQL keyword
+- Fixed a crash if non required reference fields where not given
 
 ## 0.9.9
 Released: 2016-02-11